refactor: transcripts als eigene Pipeline (analog manuals)
Statt raw/transcripts/ in den regulären Watcher-Flow zu hängen, bekommt
sie eine eigenständige Pipeline — gleiches Muster wie raw/manuals/:

- transcript-ingest.py (umbenannt von transcript-prep.py): voll integrierte
  Pipeline. Parst die Teams-.docx mit Sprecher-Awareness, schreibt eine
  strukturierte .md mit Frontmatter in raw/.cache/transcripts/<stem>.md
  und ruft anschliessend ingest.ingest() auf — der erkennt den Pfad
  raw/transcripts/ und nutzt den transkript-spezifischen Prompt-Zweig.
  Flags: --event, --date, --format, --language, --context, --force,
  --no-ingest. Sprecher-Initialen werden automatisch generiert
  (Peter Kunz -> PK; Kollisionen -> PK2/PK3).

- watch.ps1: raw/transcripts/ analog raw/manuals/ aus Startup-Scan und
  Live-Watcher ausgeschlossen. .md im Startup-Scan wieder zurueckgerollt
  (kein Bedarf mehr).

- CLAUDE.md: Verzeichnisbaum-Eintrag spiegelt manuals-Konvention
  ("eigene Pipeline: transcript-ingest.py"); watch.ps1-Doku gibt
  beide Ausnahmen an; Frontmatter-Sektion stellt klar dass Eingang
  die .docx ist (nicht Hand-.md), die .md wird vom Tool erzeugt.
